crud for categories

+++ app/Http/Controllers/CategoryController.php
@@ -19,17 +19,30 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
-        Category::create($request->validate([
-            'name' => 'required|string|max:255'
-        ]));
+        $validated = $request->validate(
+            [
+                'name' => 'required|string|max:255|unique:categories,name'
+            ]
+        );
+
+        Category::create([
+            'name' => $validated['name'],
+            'user_id' => $request->user()->id,
+            'is_default' => false,
+        ]);
+
 
         return redirect()->route('categories.index');
     }
 
     public function update(Request $request, Category $category)
     {
+        if ($category->is_default) {
+            abort(403, "This category can\'t be edited.");
+        }
+
         $category->update($request->validate([
-            'name' => 'required|string|max:255'
+            'name' => 'required|string|max:255|unique:categories.name',
         ]));
 
         return redirect()->route('categories.index');
@@ -39,6 +52,10 @@ class CategoryController extends Controller
 
     public function destroy(Category $category)
     {
+        if ($category->is_default) {
+            abort(403, "This category can\'t be deleted.");
+        }
+
         $category->delete();
         return redirect()->route('categories.index');
     }
